% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatcormatrix.R
\name{flatCorMatrix}
\alias{flatCorMatrix}
\title{flatCorMatrix}
\usage{
flatCorMatrix(matrix, pmat)
}
\arguments{
\item{matrix}{this is a matrix}

\item{pmat}{this is an output matrix}
}
\value{

}
\description{
flatCorMatrix
}
